initial commit of network model grid overhaul
I'm going to split all of the model grid functionaility that is related to nodes and links only (and is not deprecated) and put it inside of a SkeletonGrid.

the ModelGrid and the NetworkModelGrid can then both inherit from SkeletonGrid

Here skeleton is just a copy of base.py. I'll now track the changes.